PF-113: PR automation enhancements (#12)
* Prepend jira issue to PR title

- Ensures correct jira issue key(s) are prepended to PR title, based on
  PR branch name.
- Takes no action if PR title already contains issue key(s) matching
  those in branch name.
- Handles one or multiple issue keys in branch name.
- Action fails if valid issue key cannot be found.
- Action fails if PR title starts with issue key(s) that don't match
  those in branch name.

* Add PR description as comment on Jira issue

On merge of PR, add PR description as comment on Jira issue. If there
are multiple Jira issues associated with the PR, the same comment is
added to all associated Jira issues.

* Fail action if fail to create comment

If the response from Jira indicates a failure to create the comment,
or another error is caught, the GitHub Action calling this script
should exit with status 1, showing a failure status.