Ensure GH Actions scripts fail loudly
When a script called by a GH Action fails, the error should be
propagated up to the level of the GH Action itself, so that it becomes
obvious to the user.

process.exit(1) sends a 'failure' to the GH Action, causing the relevant
Action itself to fail, which the author will then get a notification
about.