Set pr_title_valid = false if no colon
For the PR title to be valid, it should start with one or more issue
keys in a valid format (as is checked in this file), and that should be
followed by a colon. This is the format expected by the Update Jira
Issue workflow, and so we should also be correctly checking for that
format here.

This commit adds checking for the colon in the rename-pr workflow.